# Conflicts:
#	pom.xml

+++ bidrag-beregn-barnebidrag/src/main/kotlin/no/nav/bidrag/beregn/barnebidrag/bo/AldersjusteringBeregningGrunnlag.kt
@@ -14,7 +14,7 @@ data class AldersjusteringBeregningGrunnlag(
     val bidragsmottakerReferanse: String,
     val bidragspliktigReferanse: String,
     val søknadsbarnPeriodeGrunnlag: SøknadsbarnPeriodeGrunnlag,
-    val vedtakId: Long,
+    val vedtakId: Int,
     val nettoTilsynsutgift: BigDecimal?,
     val barnetilsynMedStønad: BigDecimal?,
     val bpAndelFaktor: BigDecimal,
